docker changes
I had some different ideas on how docker should work.
With this setup the dependencies are installed on build, not after.
And the venv will be installed in the backend folder. So I can edit the
file there.
installing new packages require a docker rebuild I think now.

The changes here are accompanied with the changes on the backend project.